PoC/news_through_api_poc/requirements.txt
# API News Collector PoC Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
arxiv>=2.1.0
newsapi-python>=0.2.6

PyYAML>=6.0

# Optional: for better performance
requests>=2.31.0
